Use local node_modules directory to avoid dependency hell
One of girder's core packages (the datetime picker) has an upstream
depdendency (moment.js) that conflicts with an upstream dependency
on moment that is required in the older version of d3 used by the
monkeybrains plugin. We alleviate this by installing the package
locally to the plugin rather than into girder's node_modules.